fix(travel): move sync exports out of "use server" boundary
Next.js build rejected non-async exports from "use server" files
("A 'use server' file can only export async functions"). The prior
actions.ts added `MAX_STOPS_PER_TRIP` (const) and
`computeItinerarySignature` (sync function) as exports, which passed
tsc but failed the Next.js compile pass.

Move both into @/lib/travel/limits alongside MAX_RADIUS_KM (the same
file already existed precisely to hold boundary-violating exports;
see the comment at the top of actions.ts). Actions.ts now imports
them. Tests import computeItinerarySignature from @/lib/travel/limits
directly.

Also rename the signature's per-stop type to `SignatureStop` (no
`label` field) — only the subset of fields that participate in trip
identity. Removed stray `label: ""` passthroughs from callers that
were only there because SaveDestinationParams was overloaded.
